requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
importlib-metadata==4.10.0
Markdown==3.3.6
openpyxl==3.0.9
paramiko==2.9.1
Pillow==9.0.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
sshtunnel==0.4.0
transliterate==1.10.2
tzdata==2021.5
XlsxWriter==3.0.2
zipp==3.7.0
